PLANET-4811 support classic block
* Classic blocks aka "core/freeform" need to be parsed in order to
extract the same headers as would happen on the frontend. This parsing
is probably not very efficient, though the impact should only be when
there is a classic block in a post, which is discouraged.
* Move extractHeaders to a separate file.

PLANET-4811 Correct header to heading
