+++ assets/src/blocks/Submenu/SubmenuEditor.js
@@ -5,62 +5,11 @@ import { SubmenuItems } from './SubmenuItems';
 import { InspectorControls } from '@wordpress/block-editor';
 import { getSubmenuStyle } from './getSubmenuStyle';
 import { makeHierarchical } from './makeHierarchical';
-import { generateAnchor } from './generateAnchor';
+import { getHeadingsFromBlocks} from './getHeadingsFromBlocks';
 import { useSelect } from '@wordpress/data';
 
 const { __ } = wp.i18n;
 const { RichText } = wp.blockEditor;
-
-// We can put the other blocks that can have a header inside in here along with the attribute containing the heading text.
-// Then we can also filter those to include them in the menu.
-const blockTypesWithHeadings = [
-  {name: 'planet4-blocks/articles', fieldName: 'article_heading', level: 2},
-];
-
-const extractHeaders = (blocks, selectedLevels) => {
-  const headers = [];
-  blocks.forEach(block => {
-    if (block.name === 'core/heading') {
-      const blockLevel = block.attributes.level;
-
-      const levelConfig = selectedLevels.find(selected => selected.heading === blockLevel);
-
-      if (!levelConfig) {
-        return;
-      }
-
-      const anchor = block.attributes.anchor || generateAnchor(block.attributes.content);
-
-      headers.push({
-        level: blockLevel,
-        content: block.attributes.content,
-        anchor,
-        style: levelConfig.style,
-        shouldLink: levelConfig.link,
-      });
-
-      return;
-    }
-
-    const blockType = blockTypesWithHeadings.find(({ name }) => name === block.name);
-
-    if (blockType) {
-      const { fieldName, level } = blockType;
-      const levelConfig = selectedLevels.find(selected => selected.heading === level);
-
-      if (!levelConfig) {
-        return;
-      }
-      headers.push({
-        level,
-        content: block.attributes[fieldName],
-      });
-    }
-  });
-
-  return headers;
-}
-
 const renderEdit = (attributes, setAttributes) => {
   function addLevel() {
     setAttributes({ levels: attributes.levels.concat({ heading: 0, link: false, style: 'none' }) });
@@ -127,9 +76,9 @@ const renderView = (attributes, setAttributes, className) => {
     return ({ blocks: select('core/editor').getBlocks() });
   });
 
-  const flatHeaders = extractHeaders(blocks, attributes.levels);
+  const flatHeadings = getHeadingsFromBlocks(blocks, attributes.levels);
 
-  const menuItems = makeHierarchical(flatHeaders);
+  const menuItems = makeHierarchical(flatHeadings);
 
   const style = getSubmenuStyle(className, attributes.submenu_style);
