+++ assets/src/blocks/Submenu/generateAnchor.js
@@ -1,2 +1,4 @@
-export const generateAnchor = text => text.toLowerCase().trim().replace(/ /g, '-')
+export const generateAnchor = (text, previousHeadings) => {
+  return text.toLowerCase().trim().replace(/ /g, '-');
+};
